Refactoring, added ident.service/controller
