+++ src/user/models/user/user.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Get, Param, Post, Put, Query, UploadedFiles, UseGuards, UseInterceptors } from '@nestjs/common';
+import { Body, Controller, Get, Param, Put, UseGuards } from '@nestjs/common';
 import { AuthGuard } from '@nestjs/passport';
 import { ApiBearerAuth, ApiExcludeEndpoint, ApiTags } from '@nestjs/swagger';
 import { GetJwt } from 'src/shared/auth/get-jwt.decorator';
@@ -9,19 +9,14 @@ import { UserService } from './user.service';
 import { UpdateStatusDto } from './dto/update-status.dto';
 import { UserRole } from 'src/shared/auth/user-role.enum';
 import { JwtPayload } from 'src/shared/auth/jwt-payload.interface';
-import { FilesInterceptor } from '@nestjs/platform-express';
-import { KycDocument } from 'src/user/services/kyc/dto/kyc.dto';
 import { CfpVotes } from './dto/cfp-votes.dto';
-import { KycResult } from '../userData/userData.service';
-import { LimitRequestService } from '../limit-request/limit-request.service';
-import { LimitRequestDto } from '../limit-request/dto/limit-request.dto';
-import { LimitRequest } from '../limit-request/limit-request.entity';
 
 @ApiTags('user')
 @Controller('user')
 export class UserController {
-  constructor(private readonly userService: UserService, private readonly limitRequestService: LimitRequestService) {}
+  constructor(private readonly userService: UserService) {}
 
+  // --- USER --- //
   @Get()
   @ApiBearerAuth()
   @UseGuards(AuthGuard(), new RoleGuard(UserRole.USER))
@@ -36,13 +31,6 @@ export class UserController {
     return this.userService.getUser(jwt.id, true);
   }
 
-  @Get('ref')
-  @ApiBearerAuth()
-  @UseGuards(AuthGuard(), new RoleGuard(UserRole.USER))
-  async getUserRefData(@GetJwt() jwt: JwtPayload): Promise<any> {
-    return this.userService.getRefDataForId(jwt.id);
-  }
-
   @Put()
   @ApiBearerAuth()
   @UseGuards(AuthGuard(), new RoleGuard(UserRole.USER))
@@ -50,41 +38,19 @@ export class UserController {
     return this.userService.updateUser(jwt.id, newUser);
   }
 
+  // --- REF --- //
+  @Get('ref')
+  @ApiBearerAuth()
+  @UseGuards(AuthGuard(), new RoleGuard(UserRole.USER))
+  async getUserRefData(@GetJwt() jwt: JwtPayload): Promise<any> {
+    return this.userService.getRefDataForId(jwt.id);
+  }
+
   @Put('ref')
   @ApiBearerAuth()
   @UseGuards(AuthGuard(), new RoleGuard(UserRole.USER))
   async updateRef(@GetJwt() jwt: JwtPayload, @Body() { fee }: { fee: number }): Promise<number> {
     return this.userService.updateRefFee(jwt.id, fee);
-  }
-
-  @Get('kyc')
-  async getKycProgress(@Query('code') code: string): Promise<KycResult> {
-    return await this.userService.getKycProgress(code);
-  }
-
-  @Post('kyc')
-  @ApiBearerAuth()
-  @UseGuards(AuthGuard(), new RoleGuard(UserRole.USER))
-  async requestKyc(@GetJwt() jwt: JwtPayload): Promise<string> {
-    return await this.userService.requestKyc(jwt.id).then(JSON.stringify);
-  }
-
-  @Post('limit')
-  @ApiBearerAuth()
-  @UseGuards(AuthGuard(), new RoleGuard(UserRole.USER))
-  async increaseLimit(@GetJwt() jwt: JwtPayload, @Body() request: LimitRequestDto): Promise<LimitRequest> {
-    return await this.limitRequestService.increaseLimit(jwt.id, request);
-  }
-
-  @Post('incorporationCertificate')
-  @ApiBearerAuth()
-  @UseGuards(AuthGuard(), new RoleGuard(UserRole.USER))
-  @UseInterceptors(FilesInterceptor('files'))
-  async uploadIncorporationCertificate(
-    @GetJwt() jwt: JwtPayload,
-    @UploadedFiles() files: Express.Multer.File[],
-  ): Promise<boolean> {
-    return this.userService.uploadDocument(jwt.id, files[0], KycDocument.INCORPORATION_CERTIFICATE);
   }
 
   // --- CFP VOTING --- //
