+++ src/user/models/userData/userData.controller.ts
@@ -4,11 +4,12 @@ import { RoleGuard } from 'src/shared/auth/role.guard';
 import { UpdateUserDataDto } from './dto/update-userData.dto';
 import { AuthGuard } from '@nestjs/passport';
 import { UserRole } from 'src/shared/auth/user-role.enum';
-import { KycResult, UserDataService } from './userData.service';
+import { UserDataService } from './userData.service';
 import { UserData } from './userData.entity';
 import { UserDataRepository } from './userData.repository';
 import { BankDataDto } from 'src/user/models/bank-data/dto/bank-data.dto';
 import { BankDataService } from 'src/user/models/bank-data/bank-data.service';
+import { IdentService } from '../ident/ident.service';
 
 @ApiTags('userData')
 @Controller('userData')
@@ -17,6 +18,7 @@ export class UserDataController {
     private readonly userDataService: UserDataService,
     private readonly bankDataService: BankDataService,
     private readonly userDataRepo: UserDataRepository,
+    private readonly identService: IdentService,
   ) {}
 
   @Get()
@@ -43,26 +45,6 @@ export class UserDataController {
     return this.userDataRepo.findOne(id);
   }
 
-  @Put(':id/kyc')
-  @ApiBearerAuth()
-  @ApiExcludeEndpoint()
-  @UseGuards(AuthGuard(), new RoleGuard(UserRole.ADMIN))
-  async requestKyc(@Param('id') id: number): Promise<string> {
-    const userData = await this.userDataRepo.findOne({ where: { id }, relations: ['users'] });
-    const user = userData.users[0];
-    if (!user) throw new BadRequestException('UserData has no user');
-
-    return this.userDataService.requestKyc(user.id);
-  }
-
-  @Get(':id/nameCheck')
-  @ApiBearerAuth()
-  @ApiExcludeEndpoint()
-  @UseGuards(AuthGuard(), new RoleGuard(UserRole.ADMIN))
-  async getNameCheck(@Param('id') id: number): Promise<string> {
-    return this.userDataService.doNameCheck(id);
-  }
-
   @Put(':id/bankDatas')
   @ApiBearerAuth()
   @ApiExcludeEndpoint()
@@ -85,5 +67,26 @@ export class UserDataController {
   @UseGuards(AuthGuard(), new RoleGuard(UserRole.ADMIN))
   async mergeUserData(@Param('id') masterId: number, @Query('id') slaveId: number): Promise<void> {
     return this.userDataService.mergeUserData(masterId, slaveId);
+  }
+
+  // --- IDENT --- //
+  @Put(':id/kyc')
+  @ApiBearerAuth()
+  @ApiExcludeEndpoint()
+  @UseGuards(AuthGuard(), new RoleGuard(UserRole.ADMIN))
+  async requestKyc(@Param('id') id: number): Promise<string> {
+    const userData = await this.userDataRepo.findOne({ where: { id }, relations: ['users'] });
+    const user = userData.users[0];
+    if (!user) throw new BadRequestException('UserData has no user');
+
+    return this.identService.requestKyc(user.id);
+  }
+
+  @Get(':id/nameCheck')
+  @ApiBearerAuth()
+  @ApiExcludeEndpoint()
+  @UseGuards(AuthGuard(), new RoleGuard(UserRole.ADMIN))
+  async getNameCheck(@Param('id') id: number): Promise<string> {
+    return this.identService.doNameCheck(id);
   }
 }
